+++ tests/unit/spec/processors/background/VirtualBackgroundProcessor.ts
@@ -1,7 +1,7 @@
 import * as assert from 'assert';
 import * as sinon from 'sinon';
-import { BackgroundProcessor } from '../../../../lib/processors/background/BackgroundProcessor';
-import { VirtualBackgroundProcessor } from '../../../../lib/processors/background/VirtualBackgroundProcessor';
+import { BackgroundProcessor } from '../../../../../lib/processors/background/BackgroundProcessor';
+import { VirtualBackgroundProcessor } from '../../../../../lib/processors/background/VirtualBackgroundProcessor';
 
 describe('VirtualBackgroundProcessor', () => {
   let consoleWarnStub: any;
